Use github stars as metric instead of rubygems downloads
This requires you to set up a personal token and place the token in the `.oauth_token` file for use by the github api gem.

I think this is a much better metric to sort by since gen download counds for old and frequently released gems will be higher than newer and arguably more stable gems.